Move address and payment components
